+++ packages/react-zeugma/src/shared/lib/index.ts
@@ -1,4 +1,3 @@
 export * from './drag-session'
 export * from './json'
 export * from './context'
-export * from './dom'