+++ packages/react-zeugma/src/utils.ts
@@ -18,5 +18,3 @@ export {
 } from './shared/lib/tree/tree-helpers'
 
 export type { ComputedPane, ComputedSplitter } from './shared/lib/tree/tree-helpers'
-
-export { getOrCreateHiddenContainer } from './shared/lib/dom'